Created index.html
Based off of the one used for the blog section, except we decided to keep the update dates off for the view. This can be turned on later on, if necessary, but let's not focus on the dates but rather the publication. In the future we plan to add drop down categories for the publication types: essays, poems, etc.